Switching back to jQuery so I can use JSONP
The Same Origin Policy means that XhttpRequest won't work. All the
resources I'm finding online are saying that CORS doesn't work with
TMDB, so I have to use JSONP, but it seems that the only way to use that
is with jQuery.  The more you know!
Right now I'm just testing to see that connecting to TMDB doesn't cause
any other issues. The results don't print yet.